fix cbmc mode
